Implement Color Mode Data section

+++ src/main/kotlin/eu/iamgio/ktpsd/parser/PsdReader.kt
@@ -8,15 +8,15 @@ class PsdReader(private val bytes: ByteArray) {
 
     var byteIndex = 0
 
-    private fun readNext(length: Byte): ByteArray {
+    private fun readNext(length: Int): ByteArray {
         val currentBytes = bytes.slice(IntRange(byteIndex, byteIndex + length - 1)).toByteArray()
         byteIndex += length
         return currentBytes
     }
 
-    fun readNextString(length: Byte) = String(readNext(length))
+    fun readNextString(length: Int) = String(readNext(length))
 
-    private fun readNextNumber(length: Byte): Number {
+    private fun readNextNumber(length: Int): Number {
         val bytes = readNext(length)
         var value = 0
         bytes.filter {it > 0}.forEach {
@@ -25,9 +25,9 @@ class PsdReader(private val bytes: ByteArray) {
         return value
     }
 
-    fun readNextInt(length: Byte) = readNextNumber(length).toInt()
-    fun readNextByte(length: Byte) = readNextNumber(length).toByte()
-    fun readNextShort(length: Byte) = readNextNumber(length).toShort()
+    fun readNextInt(length: Int) = readNextNumber(length).toInt()
+    fun readNextByte(length: Int) = readNextNumber(length).toByte()
+    fun readNextShort(length: Int) = readNextNumber(length).toShort()
 }
 
 /**
